Architect/openevolve/pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openevolve"
dynamic = ["version"]
description = "Open-source implementation of AlphaEvolve"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "Apache-2.0"}
authors = [
    {name = "codelion"}
]
dependencies = [
    "openai>=1.0.0",
    "pyyaml>=6.0",
    "numpy>=1.22.0",
    "tqdm>=4.64.0",
    "flask",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=22.0.0",
    "isort>=5.10.0",
    "mypy>=0.950",
    "requests>=2.28.0",
]

[tool.black]
line-length = 100
target-version = ['py310']
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[project.scripts]
openevolve-run = "openevolve.cli:main"

[tool.pytest.ini_options]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests requiring external services"
]
addopts = "--strict-markers"

[tool.setuptools.packages.find]
include = ["openevolve*"]

[tool.setuptools.package-data]
openevolve = ["prompts/defaults/*.txt", "prompts/defaults/*.json"]

[tool.setuptools.dynamic]
version = {attr = "openevolve._version.__version__"}
